requirements.txt
altgraph==0.17.2
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.2
future==0.18.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
macholib==1.15.2
MarkupSafe==3.0.3
nltk==3.9.1
pymongo==4.15.3
python-dotenv==1.1.1
regex==2025.9.18
six==1.15.0
tqdm==4.67.1
Werkzeug==3.1.3
zipp==3.23.0
